feat: Update Code
- Update plugin version
- Update code structure
