+++ ROADMAP.md
@@ -2,6 +2,6 @@
 
 1. Make possible to choose default senders
 2. Make message template configurable
-3. Make Resmine import (getting data for template from issue directly)
+3. Make Redmine import (getting data for template from issue directly)
 4. Send Work Report Status Change
