+++ src/assets/inango_icon.ts
