+++ src/callbacks/on_send_reports.ts
@@ -1,0 +1,16 @@
+import sendDailyReports from "@/actions/send_daily_reports";
+import IDailyReportPlugin from "@/interfaces/plugin";
+import { Notice } from "obsidian";
+
+export default async function onSendReports(plugin: IDailyReportPlugin) {
+	new Notice('Sending Daily Reports...');
+
+	sendDailyReports(plugin)
+		.then(() => {
+			new Notice('Daily Reports Sent!');
+		})
+		.catch((err: Error) => {
+			new Notice('Error Sending Daily Reports 🫣');
+			console.error(err);
+		});
+}
