--- src/interfaces/mailer.ts
@@ -1,4 +1,0 @@
-// Mailer — interface for sending emails
-export interface Mailer {
-	send(text: string): void;
-}
