+++ src/interfaces/plugin/index.ts
@@ -1,0 +1,8 @@
+import { Plugin } from "obsidian";
+import { Settings } from "../settings";
+
+export default interface IDailyReportPlugin extends Plugin {
+	settings: Settings;
+	onload(): Promise<void>;
+	onunload(): void;
+}
