--- src/interfaces/settings/general_settings.ts
@@ -1,5 +1,0 @@
-// General settings — main settings for the plugin
-export interface GeneralSettings {
-	reportsFolder: string;
-	template: string;
-}
