+++ src/interfaces/smtp/index.ts
@@ -1,0 +1,4 @@
+// Mailer — interface for sending emails
+export interface SMTPClient {
+	send(text: string): Promise<void>;
+}
