+++ versions.json
@@ -1,3 +1,4 @@
 {
-	"0.0.1": "0.15.0"
+	"0.0.1": "0.15.0",
+	"1.0.0": "0.15.0"
 }
